refactored FriendsPage component
