+++ modules/FriendsPage/FriendsPage.module.css
@@ -1,0 +1,7 @@
+.loaders-wrapper {
+  @apply grid grid-cols-2 md:grid-cols-3 gap-3 relative;
+}
+
+.loader {
+  @apply bg-grey-bg shadow-soft p-4 rounded-md mb-xs;
+}
