+++ pages/friends/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { FC, ReactElement } from 'react';
-import FriendsPage from '@modules/HomePage/pages/FriendsPage/FriendsPage';
+import FriendsPage from '@modules/FriendsPage/FriendsPage';
 
 const index: FC = (): ReactElement => {
   return <FriendsPage />;
